Make surround downmix more similar to PHT
...because users seem to want it.

While I'm not perfectly sure that the behavior matches now, the
downmixing coefficients are now the same as with PHT. It's up to the
users to decide whether this change does any good.

We forcibly disable codec downmix, because if the user selects stereo
output, the decoder will normally do the downmix, and ignore our custom
remix options (as libswresample won't be used at all). Potentially this
has bad effects on RPI performance, or other weak CPUs.