Separate saving the settings config file
Instead of having a single save() method with a bool parameter, split it
into 2 separate functions. This doesn't really increase code
duplication, as the code was already pretty much duplicated, and makes
it clearer what is happening on the caller site.

Add a saveSection() function, which decides which config file should be
written depending on which file the section belongs to.