Use exclusive audio by default on OSX
Should solve the USB audio issues. But is equally likely to cause entire
new classes of problems, as this mode is not well-tested. On the other
hand, this uses the same OSX API as PHT, so it might be easier to match
user expectations.

With non-exclusive mode, the change-physical-format option ruins
everything for us if the OSX USB audio driver is used. It's a pretty
nasty issue that has a low chance of getting fixed properly. As we now
rely on exclusive mode for multichannel PCM, this option is not really
needed anymore, and we remove it to avoid further problems, should the
user select non-exclusive mode.