Make CEC handle properly longpress, part of #18
on some TVs the duration parameters is wrong (ie stop on Sony). In that case, it would lead to a long press, which is not handled by web.
This commit adds a predefinition of the long press to predetermine if they are supported on keys. if not, a long press will default to a short press anyways.