Fix issues with optical device type
Do not force a non-stereo channel layout in libmpv. mpv normally does
not force the channel count if passthrough is used, but the AC3
transcoding filter confuses it. (mpv does not bother to check this case;
it only looks at the decoder output - this it works only for real
passthrough.)

Hide the channel selector in the settings UI if this device type is
enabled. It only confuses everyone.