Improve InputCEC to not block the main thread.
Turns out that CEC::FindAdapters() could block the main thread and
introduce frame drops.

libcec have a new API to combat this called DetectAdapters() which
contains a quick scan. This commit changes to that API but also moves
the CEC operations to a separate thread. This is just to make sure
nothing else blocks since it's not very well documented what calls that
block and not.

Split the class into InputCEC and InputCECWorker using the same thread
model we are using for InputSDL.

Related to #129