Scale web view based on a MaxHeight
This is related to our on going problems described in #10. We now allow
web to have a max height, that means that semantic zoom will never zoom
over maxheight pixels. The rest of the scaling is done by bitmap
scaling in QML. This is all controlled by the webMaxHeight property set
in main.cpp