Remove LFE from downmixing
The LFE components are usually present in the other channels anyway. In
addition, the LFE will lower the level of other channels (due to
normalization), cause clipping if normalization is not enabled, and
introduce noise (as LFE is apparently often mastered incorrectly).

Revert it to the default, which means dropping it completely.