DisplayManagerWin: fix a little bit more and harder
So tha backwards conversion of the refresh rate in isModeMatching was
wrong. Our "fudged" refreshRate mode needs to be rounded down to get to
the DEVMODEW rate, not up.

Completely redo this: add a function to convert DEVMODEW to our own
DMVideoMode, and do the comparison always on DMVideoMode. (The use of
fabs() in modeEquals() is just a safe guard in case of relaxed float
semantics are used instead of strict IEEE semantics. Oh, and no, I'm not
implementing DMVideoMode::operator== - I'd just add more new bugs.)