Add setting to disable logging
Disabling logging was often requested by users, so here it is.

This still creates a log file at start and writes some stuff to it, even
if logging is completely disabled. The log level is set at a later stage
because I don't want to do all of the initialization (including config
file loading) without logging.