Make GL type a (hidden) setting, and enable ANGLE by default
We have much too many problems with desktop GL. Our hope is that ANGLE
will behave better.

This is messy due to interaction with QtWebEngine and the fact that it
will make QGuiApplication create/decide the GL type in the constructor.
On the other hand, most Qt stuff is not safe to be used before this
constructor is run. Our settings code definitely should run after it
(and changing it would cause an even larger mess).

To fix this, "mirror" the GL type setting as a special "use_opengl"
file, whose existence decides the GL type on init. If we make the
setting visible at a later point, the normal settings system will work,
except that it will require a restart (which needs to be done anyway).

For now, leave the setting hidden, though.