Make useOpenGL setting visible in the UI
While we're trying to avoid GUI settings that don't have much use to
normal users, this is going to be useful at least until we've solved
the remaining GL vs. ANGLE problems. (Also, it would be tiring to
explain every user on the forum the otherwise relatively roundabout
way how to switch between OpenGL and ANGLE.)