Exit Helper after 3 minutes
On windows it was pretty annoying that the Helper hanged around forever
since it was never used there anyway. So now we hang around 3 minutes
after the main application quit and then exit from the helper.
Restarting the app resets the timer. When we have the possibility to
start the main app from the helper we need to check that setting before
starting the quit timer.