Disable mpv internal OSD when processing user commands
The mpv OSD is normally unused, but for some reason mpv_command_string()
has special semantics, and will let commands display OSD messages. Not
really appropriate when running inside an application like PMP. Disable
the internal OSD completely. (This does not affect subtitles.)

Also, the "use-text-osd" option is gone - it achieved the same effect,
but was actually just a hack to soften the impact of fontconfig
misbehavior.