Make certain host commands work again
For some reason, commit baa9d41 (or one of the following ones) broke
dispatching of host commands on our CI builds. (I cannot reproduce
the problem with local builds.) We get this:

2016-01-05 20:37:34 [ WARN  ] main.cpp @ 69 - "QMetaObject::invokeMethod: No such method PlayerComponent::userCommand()\nCandidates are:\n    userCommand(QString)"

On the other hand, with this commit applied, it works.

This makes not the slightest sense to me, because userCommand has the
correct signature (and the Qt log message also shows the correct
signature). The Qt documentation also assues that the signatures
changed in this commit should be equivalent, as far as the signal
code is concerned.

I can only guess that this is caused by an unclean build or possibly
a slightly broken Qt snapshot (unlikely).

Push it anyway, so I can go on.