DisplayManagerOSX: potential robustness improvements
Clear m_osxDisplayModes on reinit. This gets rid of stale state, and
also fixes a minor memory leak. (This is probably not all that critical,
because if everything else is sane, the entries that are going to be
accessed would have been overwritten anyway.)

Check and log CGGetActiveDisplayList() and CGDisplaySetDisplayMode()
return values.