Use ANGLE by default on Windows
ANGLE is still not always better, but it is certainly infinitely better
than OpenGL as a default. OpenGL tends to cause lots of glitches,
because the interaction with DWM is bad (Microsoft doesn't care about
OpenGL at all).

On the other hand, we're stuck with ANGLE running in GLES 2.0 mode, so
no advanced filtering is available. But it still requires manual
configuration, so having to enable OpenGL mode in the settings isn't
mucn of a disimprovement.