DisplayManagerWin: fix ridiculously stupid mistake
And it's completely my own fault.

This was meant to be 1e-9 (0.000000001), but it was accidentally written
as 1e9 (1000000000), which makes the check always true, and which will
make almost always pick a mode with the wrong refresh rate.

This also explains why the refresh rate was frequently incorrect, while
all other parameters like screen size were correct.

Pointed out by a friendly forum user.