Request "Low Latency" scheduling during video playback
Use AvSetMmThreadCharacteristics() to set "Low Latency" scheduling on
the video rendering thread. In tests with mpv CLI, this has shown to be
helpful, and apparently enables higher performance by disabling energy
saving features which can interfere with playback.

Because I'm afraid that this will affect the system even if video
playback is inactive and the renderer thread is idle, do this only
during playback.