DisplayManagerX11: always pretend success when switching modes
The return value isn't always accurate, apparently. There is a user
report where it appears to return failure, but switching happened
anyway. To make it worse, the return value of XRRSetCrtcConfig() is not
documented anywhere on this earth (though implied to be a boolean by the
"Status" type), so it's not sure whether it even returns success.

Mitigation for #247.