Let the host calculate the scale for the web-client
This introduces two new pieces of information to the web-client. First
the web client is loaded with initialScale=XX on the query string.
Secondly there is now a system component signal that can be used to see
when the scale should be updated: channel.system.onScaleChanged(XX)

We need to make sure that the maxHeight is calculate by splitting it by
devicePixelRatio since size() in Qt is the *real* size and not the size
/ dpr, so when calculating the different scales, make sure to use dpr
to split max height and get a scaled value.

Also trigger the scale update when switching into fullscreen and back
to windowed mode.