PlayerComponent: avoid subtitle selection issues due to stale state
Setting options in the load command will reset the options when playback
terminates. setSubtitleStream() on the other hand usually sets permanent
options, _unless_ the corresponding options have already been marker as
playback-local. Since the load command and setSubtitleStream() set
different options depending on the situation, the result is a mess.

Resolve this by resetting all relevant options in the load command. This
potentially fixes issues with the ff-sid option not being reset and
influencing what the load command does.

The same actually applies to audio streams, but to a lesser extent, as
we don't have external audio streams yet.

(Maybe it would be better not to use the "ff-" options, and do the
stream mapping manually by retrieving the list of streams on start.)