Fix fullscreen resolution on windows at startup.
In the real world we would get a resizeEvent shortly after startup that
would resize the content area of the webview and cause some ugly black
bars around the content. This only happened when we started the app in
fullscreen. I tracked the event basically coming from system event bus,
so something is not playing nicely with Qt here. This hack should work
fine everywhere.