Fix resolution when returning from fullscreen on OSX
Before these lines when return from fullscreen on OSX it always
returned to a window that was as big as the fullscreen. That was
because on windows we had to instruct it to scale the content to
max width. On OSX we need to instruct it what the size should be
when we return to windowed mode instead.

Did I say that I love cross-platform stuff and fullscreen. It's great.