Refactor the handling of button presses and handle autorepeat internally
This commit refactors the way we handle button presses in our input
component, we now expect a button down and button up event. This is so
that we can internally handle auto-repeat instead of relying on the
input to handle it themselves.

This also lays the ground for long press detection and sequences.