Add a way to select a compatible hardware decoding mode on OSX
This changes the "hardware_decoding" setting from bool to a selection.
This breaks/discards the old setting. (I had an idea to keep it
compatible, but it seems web-client or maybe something else doesn't
like it at all if a setting in the user's config has the wrong
type, and the UI will basically become unfunctional.)

If users had a "videotoolbox-format" option set in their mpv.conf file,
it will be overwritten. (Which might be annoying for users who needed
this - they will have to explicitly select the mode in the UI after
updating.)

Actually applying the format requires a restart for silly technical
reasons. This can probably be improved later on.