Fix custom keymaps
Since the "multiple actions per pattern" change we also allowed
multiple sources to match. This broke custom keymaps since it
sent muliple actions, one for the built-in keymap and one for the
custom. We fixed this by adding a flag to CachedRegexMatcher that
makes it possible to overwrite an pattern entry.