Add option to hide power options from the exit dialog.
Users requested a way to disable the power options on their machines
since they have smaller kids or clueless people that might power
down or suspend the machine unintentionally. I opted to make this a
hidden option since it's mostly for power users.

This required a small refactoring since we wanted a central place
to check that setting instead of in every sub power component. I
chose to use the capabilites flags instead of doing the several
different methods, this seemed cleaner for the subclasses.