Allow keyinput to be mapped to a list of actions.
Since the whole mapping is a JSON object there was no way to map a single key to multiple actions
since one would always be overwritten (keys are unique) this commit adds the possibility to map a
single key to multiple actions by allowing the value of the mapping to be a list of strings. This
can't be used for more complex long/short mappings just a single action list.