PlayerQuickItem: do not create IDirect3DDevice9 in ANGLE mode
AA_UseOpenGLES indicates ANGLE is used (or the request that ANGLE should
be used - we set it in main.cpp). ANGLE does not use exclusive
fullscreen mode (unlike many drivers do in native OpenGL), so there is
no need to create a IDirect3DDevice9 in advance. (Without this, libmpv
will not be able to initialize a hardware decoder if PMP is fullscreened
and the driver uses exclusive mode.)

This is mostly untested, but should work.