Update README
Qt 5.6.0 is now out of beta, so remove any mentions of "beta".

We have removed the Qt patches from this repository. I'm not quite sure
where we publish these patches, but none of them are not required
anymore at least on Linux.

Mention that you can install a pre-built Qt.

Link to a better list of mpv dependencies.