Probe ANGLE GLES version
The tricky part here is that ANGLE does not provide a way to do this
automatically. If you request 2, you get 2. If you request 3, and 3
can't be provided, it will fail. So we have to probe this manually.

This implies Qt has been patched to pass --disable-gpu to chromium.
(Because on Windows, chromium/QtWebEngine ignores the command line it is
given.)

Using GLES 3.x allows us to do higher quality video rendering. (Though
this also has to be manually enabled, and we don't provide an automatic
setting yet.)

Unfortunately, probing for ANGLE on Windows and EGL on Linux is very
different. For ANGLE, we actually want to try create a GL context with
the specified version, so this must happen _after_ Qt loads the platform
plugin. For EGL, we want to control which platform plugin gets loaded,
so we obviously have to do before that.

The OSX-specific code could run at any point, at least as long as we
pass --disable-gpu to QtWebEngine.

This also sets QML_USE_GLYPHCACHE_WORKAROUND, which is needed to make
the debug overlay render correctly on ES 3. We expect that Qt will make
a proper fix eventually.