PlayerComponent: cleanup playback states
This goes in hand with the C++ host branch vlang-playbackstates. Neither
the C++ nor the web change are backwards-compatible.

This also gets rid of a bunch of crud that is either unnecessary, or
was never used. For example, passing the media URL via the playback
state signals was intended for gapless audio, but will have to be
redone in a better way anyway.

Fixes #485.

We have to map multiple independent states to a single state. Moreover,
these independent state changes can be sent to us in arbitrary order. We
also want to avoid sending temporary states resulting from this (e.g.
player might send us pause=false and playbackactive=true in arbitrary
order - if the latter comes first, we'd have to go to a buffering state,
and immediately leave it again). Solve this by "collecting" all state
sent from mpv until the current batch of updates stops, and then
redeciding the canonical web state.

Certain circumstances, like the single-threaded QtQuick render loop on
Windows, make this a bit harder again, and I see myself forced to create
a bunch of fields for caching state (like m_paused).