Avoid a fullscreen copy on Windows/ANGLE
Request native surface orientation from Qt, which will do the same from
ANGLE. This allows ANGLE to skip a fullscreen copy on each frame, which
is done only to invert the coordinate system. It should give us some
speedup.

As a consequence, we also have to flip our own OpenGL rendering if the
coordinate system is inverted.

This relies on a custom Qt patch, and it's not clear whether it's going
to be upstreamed. It will still build even if the patch is not applied,
because the API extension is detected at configure time.