PowerComponent: try not to use the non-existent OS screensaver on Embedded
When we enable the OS screensaver, web-client disables its own
screensaver. But Embedded has no OS screensaver, so no screen-saving
happens at all.

This is due to the "preventSystemScreensaver" setting being false. It
really should default to true on Embedded. It's already a bit too late
to change that, since there are multiple user installs with that setting
in the config file. Since setting it to false is never useful on
Embedded, just force it to true in C++.