Disable Exclusive Mode by default on OSX, allow all layouts by default
This is a change with complex consequences, but it affects only new
installations (i.e. config file defaults).

Disable using exclusive mode by default. It was supposed to be a good
thing for playing PCM surround audio out of the box. It worked, but
exclusive mode uses a weird semi-deprecated low level API on OSX, which
exposes bugs in mpv code as well audio driver code, and also causes
weird interactions with for example system volume settings.

Instead of relying on the wacky exclusive mode, disable it, and require
the user to configure the audio layout in "Audio MIDI Settings.app". I'm
not sure if there's a way to configure this automatically - I haven't
found one, and it's probably better not to mess with it automatically
anyway.

As another change, use "auto" for the channel layout setting. Since OSX
supports only two layouts (stereo and 1 multichannel layout), this
should be safe. This change will make surround PCM work automatically as
soon as the user changes the system multichannel layout in the
aforementioned "Audio MIDI Settings.app".