Revert "Use new RPI rendering method"
This partially reverts commit bb530260d8627208681f2cc03b6fec1381d0289b.

It restores the behavior on RPI before that commit. Though the code is a
bit different: the debug.force_vo option can now be set to "opengl-cb"
to test the new rendering code. We hope to use the new rendering code as
soon as the performance issues have been analyzed.

This commit is actually untested on RPI, so I'm hoping for the best.