PlayerComponent: avoid blank entry in settings if audio device missing
If an audio device was removed (due to hot plug, e.g. HDMI cable
disconnected), don't show a blank entry. Instead, show a dummy entry
that informs the user that the device is disconnected. It's also
possible to select another device and reselect the dummy one, as long as
the audio device list isn't updated by another hotplug event.

This should be much nicer to users than the weird blank entry.