OSX: hide mouse cursor harder, so it stays hidden on notifications
OSX notifications, like triggered with

    osascript -e 'display notification "hi"'

made the mouse cursor visible again. This happens behind Qt's back. Qt
most likely sets a transparent or 0-sized image, instead of actually
hiding the cursor. Work this around by using Cocoa API to hide the
cursor.

We could also just set a new blank cursor image in regular intervals,
but I think this would be a worse solution.

Sometimes the cursor fails to hide. This appears to be another Cocoa
issue, as we call [NSCursor hide] correctly. Also, the cursor is hidden
after another hide/unhide sequence, so it's not an issue with the PMP
logic, but probably a Cocoa bug.