Detection autorepeated keys manually on X11
On X11, isAutoRepeat() always returns false. This is a Qt bug, see:
https://bugreports.qt.io/browse/QTBUG-57335

There is a suggestion how to hack around this on SO:
http://stackoverflow.com/questions/38102221/qkeyevent-isautorepeat-not-working
but it requires linking to/having headers of yet another library, and
apparently doesn't fix the issue fully.

So add some lame manual detection to work this around. We simply ignore
KeyDown events until the next KeyUp event.