KonvergoWindow: add forceFSScreen setting to force display
Setting it will fullscreen PMP and/or move it to the selected screen.
Also, if the selected screen is not plugged in, plugging it in will move
the PMP window.

This is an alternative to the currently half-working old way of putting
PMP on a specific screen by windowing it, dragging it there, and then
fullscreening again.

Works on Windows, untested everywhere else.