Make more space on the debug overlay
We frequently have the problem that the debug overlay is cut off on the
bottom. QML doesn't appear to provide anything to auto-fit the text on
the debug overlay (well, maybe it shouldn't anyway). So try other means
of making more space.

Cut off long fields on the right panel (playback status). This
specifically cuts off the transcode URL at 400 characters. 400
characters should be enough to know some of the more basic info, like
source host/API and endpoint. You'll have to get the playback log to get
the full URL.

Scale the font size by the window height, not width. It was scaled by
width for the sake of long strings like log path and playback URL, but
using height seems generally better. Also use Math.round, as the
pixelSize is an integer (and will round down by default).

We could probably use pointSize, but it might be rounded too, and I'm
scared of what that might mean if hidpi is active on various platforms.